test: move CDP class out of testlib.py
