+++ test/avocado/cdp.py
@@ -1,0 +1,1 @@
+../common/cdp.py
