+++ test/common/testlib.py
@@ -45,6 +45,7 @@ import fcntl
 
 import tap
 import testvm
+import cdp
 
 TEST_DIR = os.path.normpath(os.path.dirname(os.path.realpath(os.path.join(__file__, ".."))))
 BOTS_DIR = os.path.normpath(os.path.join(TEST_DIR, "..", "bots"))
@@ -94,7 +95,7 @@ class Browser:
             self.port = port
         self.default_user = "admin"
         self.label = label
-        self.cdp = CDP("C.utf8", headless)
+        self.cdp = cdp.CDP("C.utf8", headless)
         self.password = "foobar"
 
     def title(self):
@@ -849,187 +850,6 @@ some_failed = False
 def jsquote(str):
     return json.dumps(str)
 
-class CDP:
-    def __init__(self, lang=None, headless=True):
-        self.lang = lang
-        self.timeout = 60
-        self.valid = False
-        self.headless = headless
-        self._driver = None
-        self._browser = None
-        self._browser_home = None
-        self._cdp_port_lockfile = None
-
-    def invoke(self, fn, **kwargs):
-        """Call a particular CDP method such as Runtime.evaluate
-
-        Use command() for arbitrary JS code.
-        """
-        trace = opts.trace and not kwargs.get("no_trace", False)
-        try:
-            del kwargs["no_trace"]
-        except KeyError:
-            pass
-
-        cmd = fn + "(" + json.dumps(kwargs) + ")"
-
-        # frame support for Runtime.evaluate(): map frame name to
-        # executionContextId and insert into argument object; this must not be quoted
-        # see "Frame tracking" in cdp-driver.js for how this works
-        if fn == 'Runtime.evaluate' and self.cur_frame:
-            cmd = "%s, contextId: getFrameExecId(%s)%s" % (cmd[:-2], jsquote(self.cur_frame), cmd[-2:])
-
-        if trace:
-            print("-> " + kwargs.get('trace', cmd))
-
-        # avoid having to write the "client." prefix everywhere
-        cmd = "client." + cmd
-        res = self.command(cmd)
-        if trace:
-            if "result" in res:
-                print("<- " + repr(res["result"]))
-            else:
-                print("<- " + repr(res))
-        return res
-
-    def command(self, cmd):
-        if not self._driver:
-            self.start()
-        self._driver.stdin.write(cmd + "\n")
-        line = self._driver.stdout.readline()
-        if not line:
-            self.kill()
-            raise Error("CDP broken")
-        try:
-            res = json.loads(line)
-        except ValueError:
-            print(line.strip())
-            raise
-
-        if "error" in res:
-            if opts.trace:
-                print("<- raise " + res["error"])
-            raise Error(res["error"])
-        return res["result"]
-
-    def claim_port(self, port):
-        f = None
-        try:
-            f = open(os.path.join(tempfile.gettempdir(), ".cdp-%i.lock" % port), "w")
-            fcntl.flock(f, fcntl.LOCK_EX | fcntl.LOCK_NB)
-            self._cdp_port_lockfile = f
-            return True
-        except (IOError, OSError):
-            if f:
-                f.close()
-            return False
-
-    def find_cdp_port(self):
-        """Find an unused port and claim it through lock file"""
-
-        for retry in range(100):
-            # don't use the default CDP port 9222 to avoid interfering with running browsers
-            port = random.randint (9223, 10222)
-            if self.claim_port(port):
-                return port
-        else:
-            raise Error("unable to find free port")
-
-    def start(self):
-        environ = os.environ.copy()
-        if self.lang:
-            environ["LC_ALL"] = self.lang
-        path = os.path.dirname(__file__)
-        self.cur_frame = None
-
-        # allow attaching to external browser
-        cdp_port = None
-        if "TEST_CDP_PORT" in os.environ:
-            p = int(os.environ["TEST_CDP_PORT"])
-            if self.claim_port(p):
-                # can fail when a test starts multiple browsers; only show the first one
-                cdp_port = p
-
-        if not cdp_port:
-            # start browser on a new port
-            cdp_port = self.find_cdp_port()
-            self._browser_home = tempfile.mkdtemp()
-            environ = os.environ.copy()
-            environ["HOME"] = self._browser_home
-            environ["LC_ALL"] = "C.utf8"
-            # this might be set for the tests themselves, but we must isolate caching between tests
-            try:
-                del environ["XDG_CACHE_HOME"]
-            except KeyError:
-                pass
-
-            exe = browser_path(self.headless)
-
-            if self.headless:
-                argv = [exe,  "--headless"]
-            else:
-                argv = [os.path.join(TEST_DIR, "common/xvfb-wrapper"), exe]
-
-            # sandboxing does not work in Docker container
-            self._browser = subprocess.Popen(
-                argv + ["--disable-gpu", "--no-sandbox", "--remote-debugging-port=%i" % cdp_port, "about:blank"],
-                env=environ, close_fds=True)
-            sys.stderr.write("Started %s (pid %i) on port %i\n" % (exe, self._browser.pid, cdp_port))
-
-        # wait for CDP to be up
-        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        for retry in range(300):
-            try:
-                s.connect(('127.0.0.1', cdp_port))
-                break
-            except socket.error:
-                time.sleep(0.1)
-        else:
-            raise Error('timed out waiting for browser to start')
-
-        # now start the driver
-        if opts.trace:
-            # enable frame/execution context debugging if tracing is on
-            environ["TEST_CDP_DEBUG"] = "1"
-        self._driver = subprocess.Popen(["%s/cdp-driver.js" % path, str(cdp_port)],
-                                        env=environ,
-                                        stdout=subprocess.PIPE,
-                                        stdin=subprocess.PIPE,
-                                        close_fds=True)
-        self.valid = True
-
-        for inject in [ "%s/test-functions.js" % path, "%s/sizzle.js" % path ]:
-            with open(inject) as f:
-                src = f.read()
-            # HACK: injecting sizzle fails on missing `document` in assert()
-            src = src.replace('function assert( fn ) {', 'function assert( fn ) { return true;')
-            self.invoke("Page.addScriptToEvaluateOnLoad", scriptSource=src, no_trace=True)
-
-    def kill(self):
-        self.valid = False
-        self.cur_frame = None
-        if self._driver:
-            self._driver.stdin.close()
-            self._driver.wait()
-            self._driver = None
-
-        if self._browser:
-            sys.stderr.write("Killing browser (pid %i)\n" % self._browser.pid)
-            try:
-                self._browser.terminate()
-            except OSError:
-                pass  # ignore if it crashed for some reason
-            self._browser.wait()
-            self._browser = None
-            shutil.rmtree(self._browser_home, ignore_errors=True)
-            os.remove(self._cdp_port_lockfile.name)
-            self._cdp_port_lockfile.close()
-
-    def set_frame(self, frame):
-        self.cur_frame = frame
-        if opts.trace:
-            print("-> switch to frame %s" % frame)
-
 def skipImage(reason, *args):
     if testvm.DEFAULT_IMAGE in args:
         return unittest.skip("{0}: {1}".format(testvm.DEFAULT_IMAGE, reason))
@@ -1260,32 +1080,6 @@ def arg_parser():
 
     parser.set_defaults(verbosity=1, fetch=True)
     return parser
-
-
-def browser_path(headless=True):
-    """Return path to CDP browser.
-
-    Support the following locations:
-     - "chromium-browser" in $PATH (distro package)
-     - /usr/lib*/chromium-browser/headless_shell (chromium-headless RPM), if
-       headless is true
-     - node_modules/chromium/lib/chromium/chrome-linux/chrome (npm install chromium)
-
-    Exit with an error if none is found.
-    """
-    try:
-        return subprocess.check_output(["which", "chromium-browser"]).strip()
-    except subprocess.CalledProcessError:
-        if headless:
-            g = glob.glob("/usr/lib*/chromium-browser/headless_shell")
-            if g:
-                return g[0]
-
-        p = os.path.join(os.path.dirname(TEST_DIR), "node_modules/chromium/lib/chromium/chrome-linux/chrome")
-        if os.access(p, os.X_OK):
-            return p
-
-        raise
 
 
 def test_main(options=None, suite=None, attachments=None, **kwargs):
